feat: add new default file handler
Clouds ET 2 1.0.180

feat: add default file handler for files without extension or with an extension that is not associated with any internal apps. The file handler will offer the user the option to open the file with another app, open it externally within the browser, or download the file.

feat: move the file association list to the usr system folder

refactor: move explorer gestures setup to file panel module